% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudostudy1.R
\docType{data}
\name{pseudostudy1}
\alias{pseudostudy1}
\title{pseudostudy dataset 1}
\format{pseudostudy consists of 56 cases (rows) and 5 variables where
'treatment'('age') is considered to be the explanatory factor 'A'('B') and
the variables resp1, resp2. resp3 form the response vector.}
\usage{
data(pseudostudy1)
}
\description{
Example data set in a two-way layout without
effects (in terms of disribution) by factors "treatment" and "age".
In constrast to the 'pseudostudy2' data set the number of factorial
levels is low while the number of observations per factor level combination is
large.
}
\examples{
data(pseudostudy1)
}
\keyword{datasets}
