% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudostudy2.R
\docType{data}
\name{pseudostudy2}
\alias{pseudostudy2}
\title{pseudostudy dataset 2}
\format{pseudostudy consists of 107 cases (rows) and 5 variables where
'treatment'('age') is considered to be the explanatory factor 'A'('B') and
the variables resp1, resp2. resp3 form the response vector.}
\usage{
data(pseudostudy2)
}
\description{
Example data set in a two-way layout without
effects (in terms of disribution) by factors "treatment" and "age".
}
\examples{
data(pseudostudy2)
}
\keyword{datasets}
